runtime: move stage and model resolution out of the CLI
resolve_stage_model, resolve_stages, and filter_tools_by_available now
live in leviath_runtime::pipeline::resolve, so an embedding host turns
a blueprint into ResolvedStages the exact way the daemon does. Their
one config-file input, the user's default provider/model, becomes the
plain ModelDefaults value; the CLI builds it from Config and stays the
same policy-wise. Resolver tests moved over verbatim.

Part of #72.
