+++ crates/leviath-cli/src/daemon/spawn.rs
@@ -15,14 +15,14 @@ use std::sync::{Arc, Mutex as StdMutex};
 
 use bevy_ecs::entity::Entity;
 use bevy_ecs::world::World;
-use leviath_core::blueprint::{Blueprint, ModelConfig};
+use leviath_core::blueprint::Blueprint;
 use leviath_providers::Tool;
-use leviath_runtime::ProviderRegistry;
 use leviath_runtime::host::{SpawnArgs, SubAgentOp};
 use leviath_runtime::interaction_hub::InteractionHub;
 use leviath_runtime::persistence::{RunMetadata, TokenTotals};
 use leviath_runtime::pipeline::{
-    CompactionSettings, PersistWatermark, Providers, ResolvedStage, spawn_agent_seeded,
+    CompactionSettings, ModelDefaults, PersistWatermark, Providers, resolve_stages,
+    spawn_agent_seeded,
 };
 use tokio::sync::Mutex;
 use tokio::sync::mpsc::UnboundedSender;
@@ -35,100 +35,13 @@ use crate::daemon::tool_service::{AgentToolState, CliToolService};
 /// Default max sub-agent tree depth when a blueprint doesn't set one.
 const DEFAULT_SUBAGENT_DEPTH: usize = 3;
 
-/// Resolve a stage's [`ModelConfig`] to a concrete `(provider, model)` against
-/// the registered providers. Honors a `--model` override (`provider/model` or a
-/// bare `model`), otherwise picks the first listed model whose provider is
-/// registered, then falls back to the user default (when `allow_user_default`),
-/// and finally to the config's first listed entry. (Ported from the executor's
-/// inline resolution.)
-pub fn resolve_stage_model(
-    model_cfg: &ModelConfig,
-    model_override: Option<&str>,
-    config: &Config,
-    registry: &ProviderRegistry,
-) -> (String, String) {
-    let (override_provider, override_model) = match model_override {
-        Some(ov) if ov.contains('/') => {
-            let (p, m) = ov.split_once('/').unwrap();
-            (Some(p.to_string()), Some(m.to_string()))
-        }
-        Some(ov) => (None, Some(ov.to_string())),
-        None => (None, None),
-    };
-
-    // Full provider/model override wins outright.
-    if let Some(provider) = override_provider {
-        return (provider, override_model.unwrap_or_default());
+/// The user's default provider/model from `config.toml`, in the plain form the
+/// runtime's stage resolver takes.
+pub(crate) fn model_defaults(config: &Config) -> ModelDefaults {
+    ModelDefaults {
+        provider: config.default_provider.clone(),
+        model: config.default_model.clone(),
     }
-
-    // First listed model whose provider is registered.
-    for entry in &model_cfg.models {
-        if registry.has(&entry.provider) {
-            let model = override_model
-                .clone()
-                .unwrap_or_else(|| entry.model.clone());
-            return (entry.provider.clone(), model);
-        }
-    }
-
-    // Fall back to the user's default model, or finally the first listed entry.
-    user_default_model(model_cfg, override_model.as_deref(), config, registry).unwrap_or_else(
-        || {
-            (
-                model_cfg.provider().to_string(),
-                model_cfg.model().to_string(),
-            )
-        },
-    )
-}
-
-/// The user-default fallback for [`resolve_stage_model`]: `None` when the stage
-/// forbids it or no usable default exists.
-fn user_default_model(
-    model_cfg: &ModelConfig,
-    override_model: Option<&str>,
-    config: &Config,
-    registry: &ProviderRegistry,
-) -> Option<(String, String)> {
-    if !model_cfg.allow_user_default {
-        return None;
-    }
-    if let Some(model) = override_model {
-        return Some((config.default_provider.clone(), model.to_string()));
-    }
-    if let Some(default_model) = &config.default_model
-        && registry.has(&config.default_provider)
-    {
-        return Some((config.default_provider.clone(), default_model.clone()));
-    }
-    None
-}
-
-/// Resolve every stage's provider/model + effective tool set from the blueprint.
-fn resolve_stages(
-    blueprint: &Blueprint,
-    model_override: Option<&str>,
-    config: &Config,
-    registry: &ProviderRegistry,
-    all_tool_defs: &[Tool],
-) -> Vec<ResolvedStage> {
-    blueprint
-        .stages
-        .iter()
-        .map(|stage| {
-            let (provider_name, model) =
-                resolve_stage_model(&stage.model, model_override, config, registry);
-            // Empty `available_tools` exposes no tools; otherwise filter the full
-            // set by name (alias-resolved). A name matching nothing (a typo, or an
-            // MCP tool whose server isn't installed) is simply omitted.
-            let tools = filter_tools_by_available(all_tool_defs, &stage.available_tools);
-            ResolvedStage {
-                provider_name,
-                model,
-                tools,
-            }
-        })
-        .collect()
 }
 
 /// The directories scanned for an agent's Rhai script tools, in precedence order
@@ -291,23 +204,6 @@ pub(crate) fn discover_script_tools_in(
         });
     }
     (set, names, defs)
-}
-
-/// Filter `all` tool defs down to those a stage's `available_tools` names
-/// (alias-resolved). Shared by spawn-time stage resolution and the mid-run
-/// tool-service refresh so both apply Layer-1 identically.
-pub fn filter_tools_by_available(all: &[Tool], available: &[String]) -> Vec<Tool> {
-    if available.is_empty() {
-        return Vec::new();
-    }
-    all.iter()
-        .filter(|t| {
-            available
-                .iter()
-                .any(|n| leviath_tools::canonical_tool_name(n) == t.name)
-        })
-        .cloned()
-        .collect()
 }
 
 /// Discover the agent's Rhai script tools and build their `Tool`
@@ -869,7 +765,7 @@ fn build_agent_inner(
         resolve_stages(
             &blueprint,
             args.model.as_deref(),
-            config,
+            &model_defaults(config),
             registry,
             &all_tool_defs,
         )
@@ -1167,6 +1063,8 @@ fn build_agent_inner(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use leviath_core::blueprint::ModelConfig;
+    use leviath_runtime::ProviderRegistry;
     use leviath_runtime::world::PipelineWorld;
 
     /// A throwaway sub-agent op sender for tests that don't exercise the bridge.
@@ -1330,115 +1228,6 @@ mod tests {
         fn capabilities(&self, _m: &str) -> leviath_providers::ModelCapabilities {
             leviath_providers::ModelCapabilities::default()
         }
-    }
-
-    #[test]
-    fn resolve_full_override_wins() {
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("anthropic", "x")]),
-            Some("openai/gpt-5"),
-            &Config::default(),
-            &registry_with(&[]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("openai", "gpt-5"));
-    }
-
-    #[test]
-    fn resolve_first_available_model() {
-        // anthropic not registered, openai is → picks openai.
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("anthropic", "a"), ("openai", "o")]),
-            None,
-            &Config::default(),
-            &registry_with(&["openai"]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("openai", "o"));
-    }
-
-    #[test]
-    fn resolve_model_only_override_keeps_available_provider() {
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("openai", "o")]),
-            Some("gpt-override"),
-            &Config::default(),
-            &registry_with(&["openai"]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("openai", "gpt-override"));
-    }
-
-    #[test]
-    fn resolve_user_default_when_nothing_listed_available() {
-        // Listed provider "ghost" is unavailable; anthropic (the default) is.
-        let config = Config {
-            default_provider: "anthropic".to_string(),
-            default_model: Some("claude-default".to_string()),
-            ..Default::default()
-        };
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("ghost", "g")]),
-            None,
-            &config,
-            &registry_with(&["anthropic"]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("anthropic", "claude-default"));
-    }
-
-    #[test]
-    fn resolve_user_default_with_model_override() {
-        let config = Config {
-            default_provider: "anthropic".to_string(),
-            ..Default::default()
-        };
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("ghost", "g")]),
-            Some("just-a-model"),
-            &config,
-            &registry_with(&[]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("anthropic", "just-a-model"));
-    }
-
-    #[test]
-    fn resolve_user_default_provider_unavailable_falls_through() {
-        // allow_user_default, a default_model set, but the default provider isn't
-        // registered ⇒ neither user-default branch fires ⇒ last resort.
-        let config = Config {
-            default_provider: "ghost-default".to_string(),
-            default_model: Some("dm".to_string()),
-            ..Default::default()
-        };
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("ghost", "g")]),
-            None,
-            &config,
-            &registry_with(&[]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("ghost", "g"));
-    }
-
-    #[test]
-    fn resolve_last_resort_first_listed() {
-        // No override, nothing available, no usable default → first listed entry.
-        let config = Config::default(); // default_model None
-        let (p, m) = resolve_stage_model(
-            &model_cfg(vec![("ghost", "g")]),
-            None,
-            &config,
-            &registry_with(&[]),
-        );
-        assert_eq!((p.as_str(), m.as_str()), ("ghost", "g"));
-    }
-
-    #[test]
-    fn resolve_no_user_default_uses_last_resort() {
-        let mut cfg = model_cfg(vec![("ghost", "g")]);
-        cfg.allow_user_default = false; // forbid the default fallback
-        let config = Config {
-            default_model: Some("would-be-default".to_string()),
-            ..Default::default()
-        };
-        let (p, m) = resolve_stage_model(&cfg, None, &config, &registry_with(&["anthropic"]));
-        assert_eq!((p.as_str(), m.as_str()), ("ghost", "g"));
     }
 
     // ── build_agent (full spawn from a manifest) ──
@@ -2624,63 +2413,6 @@ mod tests {
         let mut map = base();
         bump_read_sensitivities(&mut map, false);
         assert_eq!(map, base(), "no grant, no change");
-    }
-
-    #[test]
-    fn resolve_stages_empty_available_tools_gets_none() {
-        let mut stage =
-            leviath_core::Stage::new("s".to_string(), model_cfg(vec![("anthropic", "m")]));
-        stage.available_tools = vec![]; // empty ⇒ no tools
-        let layout = leviath_core::layout::ContextLayout::new(vec![], 1000);
-        let bp = Blueprint::new("t".to_string(), "d".to_string(), vec![stage], layout);
-        let tools = vec![Tool {
-            name: "read_file".to_string(),
-            description: String::new(),
-            parameters: serde_json::Value::Null,
-        }];
-        let resolved = resolve_stages(
-            &bp,
-            None,
-            &Config::default(),
-            &registry_with(&["anthropic"]),
-            &tools,
-        );
-        assert!(resolved[0].tools.is_empty());
-    }
-
-    #[test]
-    fn resolve_stages_matches_by_alias_and_skips_unknown_names() {
-        // A stage names `bash` (an alias) and a not-installed MCP tool. The
-        // filter must select the canonical `shell` definition for the alias and
-        // silently omit the unknown name (no error, no panic).
-        let mut stage =
-            leviath_core::Stage::new("s".to_string(), model_cfg(vec![("anthropic", "m")]));
-        stage.available_tools = vec!["bash".to_string(), "acme__uninstalled".to_string()];
-        let layout = leviath_core::layout::ContextLayout::new(vec![], 1000);
-        let bp = Blueprint::new("t".to_string(), "d".to_string(), vec![stage], layout);
-        let tools = vec![
-            Tool {
-                name: "shell".to_string(),
-                description: String::new(),
-                parameters: serde_json::Value::Null,
-            },
-            Tool {
-                name: "read_file".to_string(),
-                description: String::new(),
-                parameters: serde_json::Value::Null,
-            },
-        ];
-        let resolved = resolve_stages(
-            &bp,
-            None,
-            &Config::default(),
-            &registry_with(&["anthropic"]),
-            &tools,
-        );
-        let selected: Vec<&str> = resolved[0].tools.iter().map(|t| t.name.as_str()).collect();
-        // `bash` resolved to `shell`; the unknown MCP name and unlisted
-        // `read_file` were both excluded.
-        assert_eq!(selected, vec!["shell"]);
     }
 
     #[tokio::test]
